如果採用，此提交將再次更新第六次作業內容，下次再次補充筆記內容
修改項目:

模組:

議題

+++ src/test.c
@@ -7,7 +7,7 @@ int main(void) {
 
   printf("%10s%10s\n", "Fahrenheit", "Celsius");
   for(float fahrenheit = 0; fahrenheit <= 212; ++fahrenheit) {
-    celsius = ((float) 5 / 9) * (fahrenheit - 32);
+    celsius = 5.0 / 9.0 * (fahrenheit - 32);
 
     printf("%+10.3f%+10.3f\n", fahrenheit, celsius);
